==> code/old_notebooks/6_add_back_replica_data.ipynb <==
# 6) Add BG dists and assignments to Modal Share data

## env

import os
import pandas as pd
import numpy as np

### Postgres connection

# Postgres
from dotenv import load_dotenv

load_dotenv()

from sqlalchemy import create_engine
from sqlalchemy.engine import URL
from sqlalchemy import text
# os.chdir('Documents/Grad School/Berkeley/Classes/CYP_204D - Multivariate Analysis in Planning/Project/204d_final_project/')
user = "postgres"
password = os.getenv('POSTGRES_PW')
host = "127.0.0.1"
port = "5432"
database = "gradschool"

pg_url = URL.create(
    drivername='postgresql',
    username='postgres',
    host='127.0.0.1',
    database='gradschool',
    password=os.getenv('POSTGRES_PW')
)

pg_engine = create_engine(pg_url)
# Test connection
try:
    with pg_engine.connect() as connection_str:
        print('Successfully connected to the PostgreSQL database')
except Exception as e:
    print(f'Sorry failed to connect: {e}')


## Load Replica Data

# df_replica = {}
# for m in ['chi','phi','nyc','sfba']:
#     print(f'{m}..')
#     df_replica[m] = pd.read_sql(f'select * from cyp204d_final_project.{m}_replica_grpd;', con=pg_connection)
temp = pd.read_sql('select * from cyp204d_final_project.sfba_replica_grpd;', con=pg_engine.connect())

dfs_replica = {}
dfs_bg2zones = {}
for m in ['chi','phi','sfba','nyc']:
    print(f'{m}..')
    try:
        with pg_engine.connect() as pg_connection:
            #pd.read_sql_table('chi_replica_grpd', con=pg_connection, schema='cyp204d_final_project')
            dfs_replica[m] = pd.read_sql(f'select * from cyp204d_final_project.{m}_replica_grpd;', con=pg_connection)
        with pg_engine.connect() as pg_connection:
            dfs_bg2zones[m] = pd.read_sql(f'select * from cyp204d_final_project.{m}_bg2zones_od_dists;', con=pg_connection)
    except Exception as e:
        print(e)

cleaned_dfs = {}
for m in ['chi','nyc','phi','sfba']:
    print(f'{m}..')
    temp = dfs_replica[m].rename(columns={'O_bg_fips':'fips12_o', 'D_bg_fips':'fips12_d'}).copy()
    temp['trip_share'] = temp['n_trips'] / temp.groupby(['fips12_o','fips12_d'])['n_trips'].transform('sum')
    temp_pivot = temp[['fips12_o','fips12_d','primary_mode','n_trips','trip_share']].sort_values(by='n_trips',ascending=False).pivot(index=['fips12_o','fips12_d'], columns='primary_mode',values=['n_trips','trip_share'])
    df_clean = temp_pivot['trip_share'].copy()
    df_clean['total_trips'] = temp_pivot['n_trips'].sum(axis=1)
    df_clean = df_clean.fillna(0)[['total_trips'] + df_clean.columns.to_list()[0:-1]]
    cleaned_dfs[m] = df_clean.copy()


# # dfs_bg2zones['chi'].set_index(['fips12_o','fips12_d']).drop('index',axis=1)
final_dataset = {}
for m in ['chi','phi','nyc','sfba']:
    print(f'{m}..')
    final_dataset[m] = (
        dfs_bg2zones[m]
        .set_index(['fips12_o','fips12_d'])
        .drop('index',axis=1)
        .merge(
            cleaned_dfs[m],
            left_index=True,
            right_index=True
        )
    )

# #dfs_bg2zones['chi']['fips12_o']
# #cleaned_dfs['nyc']
# dfs_replica['nyc']
#dfs_replica['sfba']

# (
#     pd.concat({k: x.groupby('zone_same')[['total_trips','biking','walking','on_demand_auto','private_auto','auto_passenger','public_transit']].sum() for k, x in final_dataset.items()})
for m in ['chi','phi','nyc','sfba']:
    print(f'{m}..')
    final_dataset[m].to_sql(f'{m}_final_datasets', con = pg_engine.connect(), schema='cyp204d_final_project', if_exists='replace')


for m in final_dataset.keys():
    print(f"'{m}', {final_dataset[m]['zoneID_o'].unique().shape[0]}, {final_dataset[m]['zoneID_d'].unique().shape[0]}")

## Mapping

import pickle
import geopandas as gpd
zones = {}
for m in ['chi','phi','nyc','sfba']:
    with open(f'../data/temp/zones_{m}', 'rb') as fp:
        zones[m] = gpd.GeoDataFrame(pickle.load(fp).reset_index().drop('index',axis=1).rename(columns={0:'geometry'}))



#!pip install folium
#!pip install mapclassify
# zones['chi'].reset_index().assign(temp=lambda x: x.index % 10).explore('temp')
#zones['phi'].reset_index().assign(temp=lambda x: x.index % 10).explore('temp')
# zones['nyc'].reset_index().assign(temp=lambda x: x.index % 10).explore('temp')
# zones['sfba'].reset_index().assign(temp=lambda x: x.index % 10).explore('temp')
# #zones['chi'].reset_index().assign(temp=lambda x: x.index % 10).explore('temp')

#pd.concat([zones['chi'], zones['phi'], zones['sfba'], zones['chi']]).reset_index().assign(temp=lambda x: x.index % 10).explore('temp')
for m in zones.keys():
   print(f"{m}..")
   print(f"\t{zones[m].shape[0]}")
   print(f"\t{zones[m][zones[m].area > 1000000].shape[0]}")

#zones['chi'][zones['chi'].area > 1000000]

pd.concat(zones)
